LayoutTests/imported/w3c/web-platform-tests/content-security-policy/nonce-hiding/svgscript-nonces-hidden.tentative-expected.txt

 

FAIL Reading 'nonce' content attribute and IDL attribute. assert_equals: expected Element node <script nonce="abc" id="testScript" executed="yay">
    d... but got null
FAIL Cloned node retains nonce. assert_equals: IDL attribute expected (string) "abc" but got (undefined) undefined
FAIL Cloned node retains nonce when inserted. assert_equals: expected (string) "abc" but got (undefined) undefined
FAIL Writing 'nonce' content attribute. assert_equals: expected (string) "abc" but got (undefined) undefined
PASS Writing 'nonce' IDL attribute. 
PASS Document-written script executes. 
FAIL Document-written script's nonce value. assert_equals: expected "" but got "abc"
FAIL createElement.nonce. assert_equals: expected (object) null but got (string) "abc"
FAIL createElement.setAttribute. assert_equals: Post-insertion content expected "" but got "abc"
FAIL Nonces don't leak via CSS side-channels. assert_equals: expected "none" but got "url(\"http://localhost:8800/security/resources/abe.png\")"
